+++ src/auto_scheduler/search_policy/sketch_policy_rules.h
@@ -1,0 +1,207 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+/*!
+ * \file auto_scheduler/search_policy/sketch_policy_rules.h
+ * \brief 
+ */
+
+#ifndef TVM_AUTO_SCHEDULER_SEARCH_POLICY_SKETCH_POLICY_RULES_H_
+#define TVM_AUTO_SCHEDULER_SEARCH_POLICY_SKETCH_POLICY_RULES_H_
+
+#include <tvm/auto_scheduler/loop_state.h>
+
+#include <utility>
+#include <vector>
+
+namespace tvm {
+namespace auto_scheduler {
+
+class SketchPolicyNode;
+
+/********** Sketch Generation Rule **********/
+
+/*! \brief The base class for derivation rules used in the sketch generation. */
+class SketchGenerationRule {
+ public:
+  /*! \brief Result enumeration of the condition function. */
+  enum class ConditionKind : int {
+    /*! \brief Skip this rule and continue to try the next rules. */
+    kSkip = 0,
+    /*! \brief Apply this rule and continue to try the next rules. */
+    kApply = 1,
+    /*! \brief Apply this rule and skip the rest rules. */
+    kApplyAndSkipRest = 2
+  };
+
+  /*!
+   * \brief Condition check function of this rule.
+   * \param policy The SketchPolicyNode of this rule, some information may be used during
+   * the condition checking.
+   * \param state The original state to be checked.
+   * \param stage_id The index of the stage to process this condition check.
+   * \return The condition check result of this rule.
+   */
+  virtual ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                                      int stage_id) = 0;
+
+  /*!
+   * \brief Apply function of this rule.
+   * \param policy The SketchPolicyNode of this rule, some information may be used during
+   * the rule applying.
+   * \param state The original state to apply this rule.
+   * \param stage_id The index of the next stage to apply this rule.
+   * \return The state after applying this rule, and index of the next stage.
+   */
+  virtual std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy,
+                                                   const State& state, int stage_id) const = 0;
+};
+
+/*! \brief The rule that simply skips the current stage. It returns an unchanged state and move to
+ * the next stage. */
+class RuleSkipStage : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/*! \brief The rule that inlines simple elementwise ops.
+ * \note This rule only inlines the strictly inlineable stages. Stages marked as not strictly
+ * inlineable will have a chance to try different compute at location in InitPopulation later.
+ */
+class RuleAlwaysInline : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/*! \brief The rule that performs multi-level tiling. */
+class RuleMultiLevelTiling : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/*! The rule that performs multi-level tiling and fuses later consumers. */
+class RuleMultiLevelTilingWithFusion : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+
+ private:
+  int target_stage_id;
+};
+
+/*! \brief The rule that adds a cache write stage. */
+class RuleAddCacheWrite : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/*! \brief The rule that adds rfactor stage. */
+class RuleAddRfactor : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/*! \brief The rule that deals with compute ops that perform "fake reduction" with const tensors.
+ * This kind of op comes from winograd transformation.
+ */
+class RuleSimplifyComputeWithConstTensor : public SketchGenerationRule {
+ public:
+  ConditionKind MeetCondition(const SketchPolicyNode& policy, const State& state,
+                              int stage_id) final;
+
+  std::vector<std::pair<State, int>> Apply(const SketchPolicyNode& policy, const State& state,
+                                           int stage_id) const final;
+};
+
+/********** Init Population **********/
+
+/*! \brief The base class for derivation rules used in the initial population. */
+class InitPopulationRule {
+ public:
+  /*! \brief Result enumeration of the apply function. */
+  enum class ResultKind : int { kValid = 0, kInvalid = 1 };
+
+  /*!
+   * \brief Apply function of this rule.
+   * \param policy The SketchPolicyNode of this rule, some member may get changed during the
+   * rule applying. (e.g. random number generator)
+   * \param state The state to apply this rule, update inplace.
+   * \return The result of this rule, indicate if there's any valid state generated.
+   */
+  virtual ResultKind Apply(SketchPolicyNode* policy, State* state) const = 0;
+};
+
+/*! \brief The rule that fills the incomplete SplitSteps. */
+class InitFillTileSize : public InitPopulationRule {
+ public:
+  ResultKind Apply(SketchPolicyNode* policy, State* state) const final;
+};
+
+/*! \brief The rule that randomly changes the computation location for some stages, which do not
+ * need tiling and are not strictly inlineable(e.g. data padding). */
+class InitChangeComputeLocation : public InitPopulationRule {
+ public:
+  ResultKind Apply(SketchPolicyNode* policy, State* state) const final;
+};
+
+/*! \brief The rule that annotates parallel for CPU. */
+class InitParallel : public InitPopulationRule {
+ public:
+  ResultKind Apply(SketchPolicyNode* policy, State* state) const final;
+};
+
+/*! \brief The rule that annotates unroll. */
+class InitUnroll : public InitPopulationRule {
+ public:
+  ResultKind Apply(SketchPolicyNode* policy, State* state) const final;
+};
+
+/*! \brief The rule that annotates vectorization. */
+class InitVectorization : public InitPopulationRule {
+ public:
+  ResultKind Apply(SketchPolicyNode* policy, State* state) const final;
+};
+
+}  // namespace auto_scheduler
+}  // namespace tvm
+
+#endif  // TVM_AUTO_SCHEDULER_SEARCH_POLICY_SKETCH_POLICY_RULES_H_
